requirements.txt
discord.py
riotwatcher
python-dotenv
matplotlib
ollama
groq